# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bridge.ai sync API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/signup": {
      "post": {
        "summary": "Signup",
        "operationId": "signup_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout",
        "operationId": "logout_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Logout Auth Logout Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/session": {
      "post": {
        "summary": "Extra Session",
        "description": "The extension picks up the dashboard's login in the same browser: it trades the dashboard's\ntoken for its own (extension/src/content/dashboard.ts).",
        "operationId": "extra_session_auth_session_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Me",
        "operationId": "me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/sync": {
      "post": {
        "summary": "Sync",
        "operationId": "sync_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Sync Sync Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Credentials": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "Credentials"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HourlyTopicCount": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "hour": {
            "type": "integer",
            "maximum": 23.0,
            "minimum": 0.0,
            "title": "Hour"
          },
          "topic": {
            "type": "string",
            "enum": [
              "loneliness",
              "sadness",
              "stress",
              "anxiety",
              "anger",
              "self_worth",
              "hopelessness",
              "emptiness",
              "rejection",
              "guilt_shame",
              "overwhelm",
              "fear",
              "grief",
              "jealousy",
              "frustration",
              "disappointment",
              "embarrassment",
              "confusion",
              "exhaustion",
              "happiness",
              "calm",
              "gratitude",
              "excitement",
              "pride",
              "hope",
              "affection",
              "curiosity",
              "school",
              "friends",
              "family",
              "romance",
              "body_image",
              "boredom"
            ],
            "title": "Topic"
          },
          "count": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "hour",
          "topic",
          "count"
        ],
        "title": "HourlyTopicCount"
      },
      "LoginResult": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token",
          "email"
        ],
        "title": "LoginResult"
      },
      "Me": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email"
        ],
        "title": "Me"
      },
      "PrivacyFlag": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "hour": {
            "type": "integer",
            "maximum": 23.0,
            "minimum": 0.0,
            "title": "Hour"
          },
          "site": {
            "type": "string",
            "enum": [
              "chatgpt",
              "claude",
              "characterai",
              "gemini"
            ],
            "title": "Site"
          },
          "what": {
            "type": "string",
            "enum": [
              "message",
              "file"
            ],
            "title": "What"
          },
          "findings": {
            "items": {
              "type": "string",
              "enum": [
                "phone",
                "email",
                "ssn",
                "card",
                "bank",
                "address",
                "password",
                "birthday",
                "student_id",
                "id_document",
                "whereabouts",
                "ai_relationship",
                "unsafe"
              ]
            },
            "type": "array",
            "maxItems": 13,
            "minItems": 1,
            "title": "Findings"
          },
          "sent": {
            "type": "boolean",
            "title": "Sent"
          },
          "hidden": {
            "type": "boolean",
            "title": "Hidden",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "hour",
          "site",
          "what",
          "findings",
          "sent"
        ],
        "title": "PrivacyFlag",
        "description": "Personal info caught before it was sent. The kinds of info only, never the values."
      },
      "SiteAggregate": {
        "properties": {
          "site": {
            "type": "string",
            "enum": [
              "chatgpt",
              "claude",
              "characterai",
              "gemini"
            ],
            "title": "Site"
          },
          "level": {
            "type": "string",
            "enum": [
              "healthy",
              "watch",
              "concerning",
              "crisis"
            ],
            "title": "Level"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "active_minutes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Active Minutes"
          },
          "late_night_sessions": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Late Night Sessions"
          },
          "nudges_shown": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Nudges Shown"
          },
          "privacy_pauses": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Privacy Pauses"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "site",
          "level",
          "score",
          "active_minutes",
          "late_night_sessions",
          "nudges_shown",
          "privacy_pauses"
        ],
        "title": "SiteAggregate"
      },
      "SyncPayload": {
        "properties": {
          "child_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Child Id"
          },
          "device_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Device Id"
          },
          "week_start": {
            "type": "string",
            "format": "date",
            "title": "Week Start"
          },
          "sites": {
            "items": {
              "$ref": "#/components/schemas/SiteAggregate"
            },
            "type": "array",
            "maxItems": 4,
            "title": "Sites"
          },
          "hourly_topics": {
            "items": {
              "$ref": "#/components/schemas/HourlyTopicCount"
            },
            "type": "array",
            "maxItems": 5544,
            "title": "Hourly Topics"
          },
          "privacy_flags": {
            "items": {
              "$ref": "#/components/schemas/PrivacyFlag"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Privacy Flags"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "child_id",
          "device_id",
          "week_start",
          "sites",
          "hourly_topics"
        ],
        "title": "SyncPayload",
        "description": "One device's week. Several devices (browsers) can report for the same child_id."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}